Rename event namespace
