+++ src/Event/SourceCompilation/SourceCompileEventInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Event\SourceCompile;
+namespace App\Event\SourceCompilation;
 
 use Psr\EventDispatcher\StoppableEventInterface;
 use webignition\BasilCompilerModels\OutputInterface;
